pyproject.toml
[project]
name = "ucp-merchant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]
